pyproject.toml
[project]
name = "phishing-detection-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "certifi>=2025.11.12",
    "dagshub>=0.6.3",
    "fastapi>=0.121.1",
    "from-root>=1.3.0",
    "mlflow>=3.6.0",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pyaml>=25.7.0",
    "pymongo[srv]>=4.15.4",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.7.2",
    "setuptools>=80.9.0",
    "uvicorn>=0.38.0",
]
